fix: user sign mapper 제거
